+++ docs/list_tutorials.jl
@@ -39,6 +39,7 @@ tutorials = [
     "Calibrating a ClimaLand model" => [
         "Single site perfect model" => "calibration/perfect_model_site_level_calibration.jl",
         "Single site observations" => "calibration/obs_site_level_calibration.jl",
+        "Single site perfect model (CES/UQ)" => "calibration/perfect_model_site_level_ces.jl",
     ],
     "Running coupled simulations" =>
         ["Coupled bucket model" => "standalone/Bucket/coupled_bucket.jl"],